all: implement amd64 support

+++ src/cmd/mk/mk.c
@@ -213,11 +213,11 @@ outofdate(Node *node, Arc *arc, int eval)
 				str = strdup(buf);
 			ret = pcmp(arc->prog, node->name, arc->n->name, arc->r->shellt, arc->r->shellcmd);
 			if(sym)
-				sym->value = (void *)ret;
+				sym->value = (void *)(uintptr)ret;
 			else
-				symlook(str, S_OUTOFDATE, (void *)ret);
+				symlook(str, S_OUTOFDATE, (void *)(uintptr)ret);
 		} else
-			ret = (int)sym->value;
+			ret = (int)(uintptr)sym->value;
 		return(ret-1);
 	} else if(strchr(arc->n->name, '(') && arc->n->time == 0)  /* missing archive member */
 		return 1;
