+++ src/lib9/sprint.c
@@ -25,8 +25,8 @@ sprint(char *buf, char *fmt, ...)
 	 * the stack might be near the top of memory, so
 	 * we must be sure not to overflow a 32-bit pointer.
 	 */
-	if(buf+len < buf)
-		len = -(uint)buf-1;
+	if((uintptr)buf+len < (uintptr)buf)
+		len = (uint)-(uintptr)buf-1;
 
 	va_start(args, fmt);
 	n = vsnprint(buf, len, fmt, args);
